backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Obsidian WebDev - AI-powered platform builder backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "motor>=3.7.1",
    "pymongo>=4.0.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=44.0.0",
    "pycryptodome>=3.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "slowapi>=0.1.9",
    "aiofiles>=23.0.0",
    "docker>=7.1.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "tavily-python>=0.5.0",
    "duckduckgo-search>=7.0.0",
    "pypdf>=5.0.0",
]
